Add CONTENT.md feature
This feature when enabled allows the creatin of CONTENT.md file in the
account git repository. This file is similar to CONTENT, but contains
only posts (no replies or tags or scores) and use markdown format.

The idea behind this feature is to use tl as microbloggin tool too,
specially focused on new gemini protocol microbloggin ideas.

Signed-off-by: Andrés J. Díaz <ajdiaz@ajdiaz.me>