fix!: solve security problem related with replies and labels
A malicious attack can create a post and get some engagement for it, for instance
replies, labels or scores, and then change the content of that post editing the
CONTENT file and get benefits of the previous engagement.

Of course the network will be solve the situation with time, reverting the scores
and taggin properly the spam, but still it is not a good solution.

This commit orphans the replies/labes/tag which do not match with the exact
content that they refer to. Adding an internal hmac to the post and verify
that the reply has the same referer hmac.

This change is a breaking-change. A new config value is set
("timeline.show-unstrusted-items"). When set to false the unverified
replies/labels/scores will be ignored from the timeline. If true old behaviour
is respected for compatibility.

In version >= 1.9 this option will be disable by default (i.e. only verified
replies will be used).

Signed-off-by: Andrés J. Díaz <ajdiaz@ajdiaz.me>