feat: add account preview
New command added: account preview <url>
This command allows the user to get the content of a remote account/repo
without follow it or merge it in the main account path. When tl execution
ends, the temporary directory will be removed.

Plase note that big repos could cause a performace issue and a potential
attack if /tmp dir is in memory.

Signed-off-by: Andrés J. Díaz <ajdiaz@ajdiaz.me>