feat: ensure that timeline index is regenerated on post/reply
When you post and want to tag/score/reply your last posts, you dont want to run
a tl tl again. So this patch will run timeline_index_save for you when post and
reply. So it means that replying to -1 index will result in a reply of the latest
post or in a reply of the latest reply, but a tag or score to -1 index will result
to the last post/reply, but not to the last tag/score.

This is, IMHO the usual way to work. Is not very common to tag a tag or score a
score (although its still possible, previously run a tl tl), but more often you
want to tag/score your previous post/reply.

Signed-off-by: Andrés J. Díaz <ajdiaz@ajdiaz.me>