feat: add abbreviations
Abbreviations are a way to replace some text on input post or
reply by a conventional one. For example replacing *IMHO* for
*in my humble opinnion*.

This commit also add emoji support as abbreviation. For instance
use *:love:* to put a lovely hearth in your message. You will
need to enable it via config setting *input.emojis* to true.